Remove redundant casting and improve const-correctness
Eliminate redundant static_cast in playlistcore.cpp where QRandomGenerator::bounded(T) deduces the type automatically (clang-tidy readability-redundant-casting). Mark savePlaylist() const in playlistmanager.h since the method only reads state. Both changes clarify intent and strengthen const-correctness per the functional core pattern.

+++ src/services/playlistcore.cpp
@@ -180,7 +180,7 @@ QList<int> generateShuffleOrder(int itemCount, quint32 seed)
     // Fisher-Yates shuffle with explicit seed
     QRandomGenerator rng(seed);
     for (int i = itemCount - 1; i > 0; --i) {
-        int j = static_cast<int>(rng.bounded(i + 1));
+        int j = rng.bounded(i + 1);
         std::swap(order[i], order[j]);
     }
